Bug fix and general refactoring.
Bug fix: Fixed a bug where the random scrap charge was only for the host and also when they bough items.

General refactoring, better folder structure.

Already have some classes for spawning a body, but it does not work well yet for multiplayer. So it is not yet in game now.