Creating body as scrap (failed)
Some general refactoring and a lot of attempts to make the body a good scrap. Due to the masked enemy having no colliders in joints, it could not succeed in being a ragdoll. Other attempts of using the player ragdoll failed because the item is added too late. So the body as as scrap is for now not enabled.